Move settings management to a new class
